chore: move PrintPage to its own dir
